Use size_t in LRU to allow cache sizes larger than 4gb

+++ intern/cycles/blender/blender_sync.cpp
@@ -481,7 +481,7 @@ SceneParams BlenderSync::get_scene_params(BL::Scene& b_scene,
 		params.use_qbvh = false;
 	}
 
-	params.geom_cache_max_size = get_int(cscene, "geom_cache_max_size") * 1024*1024;
+	params.geom_cache_max_size = (size_t)get_int(cscene, "geom_cache_max_size") * 1024*1024;
 
 	return params;
 }
